fix(scout-for-lol): split Explore answer schemas under max-lines
Why:
Adding `includeVisualization` pushed `reports/explore.ts` over the 500-line
lint cap, which fails `@scout-for-lol/data` lint in CI.

What:
Move the answer/wire schemas into `explore-answer.ts` and export them from
the data model barrel. Existing `@scout-for-lol/data` imports stay the same.

Verification:
- `bunx turbo run lint test typecheck --filter=@scout-for-lol/data`
- Not verified: Buildkite on this commit until push.
