+++ packages/scout-for-lol/packages/data/src/model/index.ts
@@ -30,6 +30,7 @@ export * from "./competitions/queue-window-drift.ts";
 export * from "./riot/rank.ts";
 export * from "./reports/report.ts";
 export * from "./reports/explore.ts";
+export * from "./reports/explore-answer.ts";
 export * from "./core/form-inputs.ts";
 export * from "./reports/report-ai.ts";
 export * from "./reports/report-query-champions.ts";
